LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbcursor_update_objectstore9-expected.txt


PASS IDBCursor.update() - object store - throw InvalidStateError when the cursor is being iterated
